+++ pystella/decomp.py
@@ -32,7 +32,50 @@ class DomainDecomposition:
 
     If :mod:`mpi4py` is not installed, then only single-rank operation is supported.
 
-    .. automethod:: __init__
+    :arg proc_shape: A 3-:class:`tuple` specifying the shape of the MPI
+        processor grid.
+
+        .. note::
+
+            Currently, ``proc_shape[2]`` must be ``1``, i.e., only
+            two-dimensional domain decompositions are supported.
+
+    :arg halo_shape: The number of halo layers on (both sides of) each axis of
+        the computational grid (i.e., those used for grid synchronization or
+        enforcing boundary conditions).
+        May be a 3-:class:`tuple` or an :class:`int` (which is interpreted
+        as that value for each axis).
+        Note that zero is a valid value along any axis (in which case grid
+        synchronization for such axes is skipped).
+        For example, ``halo_shape=(2,0,1)`` specifies that there are two
+        additional layers at each the beginning and end of the first axis, none
+        for the second, and one for the third.
+        For ``rank_shape=(Nx, Ny, Nz)``, an appropriately padded array would
+        then have shape ``(Nx+4, Ny, Nz+2)``.
+
+    The following keyword arguments are recognized:
+
+    :arg rank_shape: A 3-:class:`tuple` specifying the shape of the computational
+        sub-grid on the calling process.
+        Defaults to *None*, in which case the global size is not fixed (and
+        will be inferred when, e.g., :meth:`share_halos` is called, at a slight
+        performance penalty).
+
+    :arg grid_shape: A 3-:class:`tuple` specifying the shape of the global
+        computational grid.
+        If passed, :attr:`rank_shape` is determined for each rank.
+        Note that :attr:`proc_shape` need not evenly divide ``grid_shape``, i.e.,
+        different ranks may have different :attr:`rank_shape`\\ s.
+
+    :raises NotImplementedError: if ``proc_shape[2] != 1``.
+
+    :raises ValueError: if the size of the processor grid
+        ``proc_shape[0] * proc_shape[1] * proc_shape[2]`` is not equal to the
+        total number of ranks the application was launched with
+        (i.e., that returned by :func:`mpi4py.MPI.COMM_WORLD.Get_size()`).
+
+    :raises ValueError: if both ``rank_shape`` and ``grid_shape`` are passed.
+
     .. automethod:: share_halos
     .. automethod:: remove_halos
     .. automethod:: gather_array
@@ -42,6 +85,7 @@ class DomainDecomposition:
     .. autoattribute:: rank_tuple
     .. automethod:: bcast
     .. automethod:: allreduce
+    .. automethod:: get_rank_shape_start
 
     .. attribute:: comm
 
@@ -62,45 +106,7 @@ class DomainDecomposition:
     .. attribute:: rank_shape
     """
 
-    def __init__(self, proc_shape, halo_shape, rank_shape=None):
-        """
-        :arg proc_shape: A 3-:class:`tuple` specifying the shape of the MPI
-            processor grid.
-
-            .. note::
-
-                Currently, ``proc_shape[2]`` must be ``1``, i.e., only
-                two-dimensional domain decompositions are supported.
-
-        :arg halo_shape: The number of halo layers on (both sides of) each axis of
-            the computational grid (i.e., those used for grid synchronization or
-            enforcing boundary conditions).
-            May be a 3-:class:`tuple` or an :class:`int` (which is interpreted
-            as that value for each axis).
-            Note that zero is a valid value along any axis (in which case grid
-            synchronization for such axes is skipped).
-            For example, ``halo_shape=(2,0,1)`` specifies that there are two
-            additional layers at each the beginning and end of the first axis, none
-            for the second, and one for the third.
-            For ``rank_shape=(Nx, Ny, Nz)``, an appropriately padded array would
-            then have shape ``(Nx+4, Ny, Nz+2)``.
-
-        The following keyword arguments are recognized:
-
-        :arg rank_shape: A 3-:class:`tuple` specifying the shape of the computational
-            sub-grid on the calling process.
-            Defaults to *None*, in which case the global size is not fixed (and
-            will be inferred when, e.g., :meth:`share_halos` is called, at a slight
-            performance penalty).
-
-        :raises NotImplementedError: if ``proc_shape[2] != 1``.
-
-        :raises ValueError: if the size of the processor grid
-            ``proc_shape[0] * proc_shape[1] * proc_shape[2]`` is not equal to the
-            total number of ranks the application was launched with
-            (i.e., that returned by :func:`mpi4py.MPI.COMM_WORLD.Get_size()`).
-        """
-
+    def __init__(self, proc_shape, halo_shape, rank_shape=None, grid_shape=None):
         self.proc_shape = proc_shape
         self.halo_shape = ((halo_shape,)*3 if isinstance(halo_shape, int)
                            else halo_shape)
@@ -128,6 +134,11 @@ class DomainDecomposition:
         self.rz = self.rank % proc_shape[2]
         self.ry = (self.rank - self.rz) // proc_shape[2] % proc_shape[1]
         self.rx = (self.rank - self.rz - proc_shape[2] * self.ry) // proc_shape[1]
+
+        if grid_shape is not None:
+            if self.rank_shape is not None:
+                raise ValueError("Only pass one of rank_shape or grid_shape.")
+            self.rank_shape, _ = self.get_rank_shape_start(grid_shape)
 
         params_to_fix = dict(hx=self.halo_shape[0],
                              hy=self.halo_shape[1],
@@ -240,17 +251,20 @@ class DomainDecomposition:
         self.gather_knl = make_G_S_knl("arr[i, j, k] = subarr[i+hx, j+hy, k+hz]")
         self.scatter_knl = make_G_S_knl("subarr[i+hx, j+hy, k+hz] = arr[i, j, k]")
 
-    def get_displs_and_counts(self, full_shape, rank_shape):
+    def get_displs_and_counts(self, full_shape, x_slice):
         NX, NY, NZ = full_shape
-        Nx, Ny, Nz = rank_shape
 
         displs = np.ones(np.product(self.proc_shape[:2]), dtype='int')
+        counts = np.ones(np.product(self.proc_shape[:2]), dtype='int')
         for ri in range(self.proc_shape[0]):
             for rj in range(self.proc_shape[1]):
+                (Nx, Ny, Nz), (sx, sy, sz) = self.get_rank_shape_start(
+                    full_shape, (ri, rj, 0)
+                )
                 rid = self.rankID(ri, rj, 0)
-                displs[rid] = NZ * (rj * Ny + NY * ri * Nx)
-
-        counts = Nz * Ny * np.ones(np.product(self.proc_shape[:2]), dtype='int')
+                # repeat the last row when some ranks have smaller rank_shape[0]
+                displs[rid] = NZ * (sy + NY * sx) + min(x_slice, Nx-1) * NY * NZ
+                counts[rid] = Nz * Ny
 
         return displs, counts
 
@@ -272,6 +286,39 @@ class DomainDecomposition:
         """
 
         return (self.rx, self.ry, self.rz)
+
+    def get_rank_shape_start(self, grid_shape, rank_tuple=None):
+        """
+        Computes the calling process's ``rank_shape`` and starting indices
+        ``rank_start`` (as would be indexed into the global array)
+        corresponding to a global array with shape ``grid_shape``.
+
+        :arg grid_shape: A 3-:class:`tuple` specifying the shape of the
+            corresponding global array.
+
+        :arg rank_tuple: A 3-:class:`tuple` specifying for which MPI rank
+            in the processor grid for which to compute outputs.
+            Defaults to that of the calling process (i.e., :attr:`rank_tuple`).
+
+        :returns: ``(rank_shape, rank_start)``.
+        """
+
+        # borrowed from mpi4py_fft.pencil
+        def _blockdist(N, size, rank):
+            q, r = divmod(N, size)
+            n = q + (1 if r > rank else 0)
+            s = rank * q + min(rank, r)
+            return (n, s)
+
+        rank_shape = [None] * 3
+        rank_start = [None] * 3
+        rank_tuple = rank_tuple or self.rank_tuple
+        for i, (Npi, ri) in enumerate(zip(self.proc_shape, rank_tuple)):
+            n, s = _blockdist(grid_shape[i], Npi, ri)
+            rank_shape[i] = n
+            rank_start[i] = s
+
+        return tuple(rank_shape), tuple(rank_start)
 
     def _get_or_create_halo_buffers(self, queue, shape, dtype):
         if (shape, dtype) not in self.buffer_arrays:
@@ -297,7 +344,8 @@ class DomainDecomposition:
             The number of halo layers on each face of the grid is fixed by
             :attr:`halo_shape`, while the shape of ``fx``
             (i.e., subtracting halo padding specified by :attr:`halo_shape`)
-            is only fixed if ``rank_shape`` was passed to :meth:`__init__`.
+            is only fixed if ``rank_shape`` or ``grid_shape`` were
+            passed at object creation).
         """
 
         h = self.halo_shape
@@ -464,29 +512,35 @@ class DomainDecomposition:
             self.remove_halos(queue, in_array, out_array)
         else:
             temp_in = in_array.get() if cl_in else in_array
-            if cl_out:
+            if cl_out and self.rank == root:
                 temp_out = np.zeros(shape=out_array.shape, dtype=dtype)
-            else:
+            elif self.rank == root:
                 temp_out = out_array
+            else:
+                temp_out = None
 
             rank_shape = tuple(i - 2 * hi for i, hi in zip(in_array.shape, h))
             full_shape = None if self.rank != root else out_array.shape
             full_shape = self.bcast(full_shape, root)
-            displs, counts = self.get_displs_and_counts(full_shape, rank_shape)
 
             from mpi4py import MPI
             MPI_dtype = MPI._typedict[np.dtype(dtype).char]
 
-            xy_slice = tuple(slice(hi, -hi) if hi > 0 else slice(None) for hi in h)
+            interior = tuple(slice(hi, -hi) if hi > 0 else slice(None) for hi in h)
 
-            for i in range(rank_shape[0]):
-                slc = (i+h[0],) + xy_slice[1:]
-                tmp = temp_in[slc].copy()
-                self.comm.Gatherv(tmp,
-                                  [temp_out[i], counts, displs, MPI_dtype],
-                                  root=root)
+            nx = self.allreduce(rank_shape[0], op=MPI.MAX)
+            for i in range(nx):
+                displs, counts = self.get_displs_and_counts(full_shape, i)
+                # repeat the last row when some ranks have smaller rank_shape[0]
+                x_slice = min(i + h[0], rank_shape[0] + h[0] - 1)
+                tmp = temp_in[(x_slice,) + interior[1:]].copy()
+                if self.rank == root:
+                    recv_buf = [temp_out, counts, displs, MPI_dtype]
+                else:
+                    recv_buf = None
+                self.comm.Gatherv(tmp, recv_buf, root=root)
 
-            if cl_out:
+            if cl_out and self.rank == root:
                 out_array.set(temp_out)
 
     def restore_halos(self, queue, in_array, out_array):
@@ -590,22 +644,21 @@ class DomainDecomposition:
             else:
                 temp_out = out_array
 
-            displs, counts = self.get_displs_and_counts(full_shape, rank_shape)
-
             from mpi4py import MPI
             MPI_dtype = MPI._typedict[np.dtype(dtype).char]
 
-            xy_slice = tuple(slice(hi, -hi) if hi > 0 else slice(None) for hi in h)
+            interior = tuple(slice(hi, -hi) if hi > 0 else slice(None) for hi in h)
 
-            for i in range(rank_shape[0]):
-                tmp_in = None
+            nx = self.allreduce(rank_shape[0], op=MPI.MAX)
+            for i in range(nx):
+                displs, counts = self.get_displs_and_counts(full_shape, i)
                 if self.rank == root:
-                    tmp_in = temp_in[i, :, :]
-                self.comm.Scatterv([tmp_in, counts, displs, MPI_dtype],
-                                   tmp,
-                                   root=root)
-                slc = (i+h[0],) + xy_slice[1:]
-                temp_out[slc] = tmp
+                    send_buf = [temp_in, counts, displs, MPI_dtype]
+                else:
+                    send_buf = None
+                self.comm.Scatterv(send_buf, tmp, root=root)
+                if i < rank_shape[0]:
+                    temp_out[(i + h[0],) + interior[1:]] = tmp
 
             if cl_out:
                 out_array.set(temp_out)