+++ doc/changes.rst
@@ -12,6 +12,8 @@ Version 2020.2
 
 * Low-storage Runge-Kutta timesteppers handle temporary arrays internally
   and no longer constrain all degrees of freedom to have identical datatypes and shapes.
+* Added support for :class:`DomainDecomposition`\ s
+  with processor grid dimensions that do not evenly divide the global computational grid.
 
 Version 2020.1
 --------------